Created Turret Type Selector UI
I created a second turret node with a new skin imported as a glb file. I used a SubViewPortContainer and SubViewport to generate images of each tower that could be used as icons for the tower type selection UI. The UI contains an Hbox with two buttons that will contain logic that controls what type of tower is placed in the future.